// src/app/(authed)/communities/schemas/build-community-schema.ts
import { z } from "zod";

export const buildCommunitySchema = z.object({
  name: z.string().min(1, "Name is required"),
  email: z.string().min(1, "Email is required").email("Invalid email address"),
  message: z.string().min(1, "Message is required"),
});

export type IBuildCommunitySchema = z.infer<typeof buildCommunitySchema>;